Integrate forms and functions
Integrate the forms and functions responsible for
updating the library and displaying books on screen.

Create a new and unique div to display the books.

There was a bug where the function 'displayBooks()'
duplicated the results every time we called it.
To fix this problem, we added a function to clean the
new div and avoid duplicates.

fix typo: 'read' to 'readOrNot'.