Fix 'removeBook()' function
The function has to remove a book object
from two places, the screen and the array.

The removal starts from the screen when the
user presses a button to remove the selected
book.

Doing so, the data-attribute of each remaining
book has to change. It's important because if
the data-attribute does not match the book index
in the array, bugs start to happen.

bug example: If you remove a book of
data-book-index '01', you are removing the book
object at index 01 from the array. When the
data-attribute does not change for the remaining
books, the book of data-book-index '02' targets
a different object in the array.

The previous version of the function did not
cover these problems. Now, it's fixed.