Style the book cards
Adjust the size and position of the book cards.

Use flexbox to properly stack the cards.

Add a pink and green color.

Add a gradient background color to the title
section.

Invert the position of 'author' and 'title'
elements on the HTML file. Do the same for
the buttons responsible for removing the book
or change it's reading status.