Fix 'removeBook()' function
Somehow, there was a bug where if all cards
were removed and depending on the deletion
order, when the user tried to submit a new
book, one of the last deleted objects would
be displayed on the screen again, next to the
new book.

To fix that, now the function checks if the
parent container has children or nothing is
displayed on the screen. If that's the case,
the array responsible for storing the books is
cleaned.

Remove some unnecessary 'console.log' calls.